extract Graph component to its own file
