+++ frontend/src/components/GraphVisualization/index.tsx
@@ -1,5 +1,4 @@
 import { useState } from 'react'
-import CytoscapeComponent from 'react-cytoscapejs'
 import {
   AddIcon,
   DeleteIcon,
@@ -31,8 +30,6 @@ import {
   AutoCompleteItem,
   AutoCompleteList,
 } from '@choc-ui/chakra-autocomplete'
-import cytoscape from 'cytoscape'
-import cola from 'cytoscape-cola'
 import { useExecutionStore } from 'stores'
 
 import {
@@ -42,7 +39,7 @@ import {
   assertAdjacencyMatrix,
 } from 'utils/graph'
 
-cytoscape.use(cola)
+import { Graph } from './Graph'
 
 enum EdgeFormat {
   AdjacencyMatrix = 'Adjacency Matrix',
@@ -278,94 +275,5 @@ export const GraphVisualization = (props: GraphVisualizationProps) => {
         </ModalContent>
       </Modal>
     </Box>
-  )
-}
-
-interface GraphProps {
-  directed: boolean
-  weighted: boolean
-  adjacencyList: number[][] | [number, number][][]
-  displayData: { name: string; array: string[] }[]
-}
-
-const Graph = (props: GraphProps) => {
-  const { directed, weighted, adjacencyList, displayData } = props
-
-  const n = adjacencyList.length
-  const elements: cytoscape.ElementDefinition[] = []
-  for (let i = 0; i < n; i++) {
-    const label = i.toString()
-    const contents = []
-    contents.push(label)
-    for (const { name, array } of displayData) {
-      if (array.length <= i) continue
-      contents.push(`${name}: ${array[i]}`)
-    }
-    elements.push({
-      group: 'nodes',
-      data: {
-        id: i.toString(),
-        label,
-        content: contents.join('\n'),
-      },
-    })
-  }
-
-  for (let i = 0; i < n; i++) {
-    if (weighted) {
-      for (const [v, weight] of adjacencyList[i] as [number, number][]) {
-        elements.push({
-          group: 'edges',
-          data: {
-            source: i.toString(),
-            target: v.toString(),
-            label: weight.toString(),
-          },
-        })
-      }
-    } else {
-      for (const v of adjacencyList[i] as number[]) {
-        elements.push({
-          group: 'edges',
-          data: {
-            source: i.toString(),
-            target: v.toString(),
-          },
-        })
-      }
-    }
-  }
-
-  return (
-    <CytoscapeComponent
-      elements={elements}
-      layout={{ name: 'cola' }}
-      style={{ height: '400px', width: '100%' }}
-      stylesheet={[
-        {
-          selector: 'edge',
-          style: {
-            width: 4,
-            ...(directed && {
-              'target-arrow-shape': 'triangle',
-              'curve-style': 'bezier',
-            }),
-            ...(weighted && {
-              label: 'data(label)',
-            }),
-          },
-        },
-        {
-          selector: 'node',
-          style: {
-            'border-color': 'black',
-            'border-width': '1px',
-            label: 'data(content)',
-            'font-size': '0.75em',
-            'text-wrap': 'wrap',
-          },
-        },
-      ]}
-    />
   )
 }
